+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `alias` keyWord location to Parsed Participant Definition
 
 ## 1.0.0 - 2020-11-05
 - Initial implementation